+++ src/main/java/com/rualone/app/domain/board/application/PostService.java
@@ -1,6 +1,6 @@
 package com.rualone.app.domain.board.application;
 
-import com.rualone.app.domain.board.Post;
+import com.rualone.app.domain.board.entity.Post;
 import com.rualone.app.domain.board.dto.request.PostCreateRequest;
 import com.rualone.app.domain.board.dto.request.PostUpdateRequest;
