+++ src/main/java/com/rualone/app/domain/board/dao/CommentRepository.java
@@ -1,6 +1,6 @@
 package com.rualone.app.domain.board.dao;
 
-import com.rualone.app.domain.board.Comment;
+import com.rualone.app.domain.board.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
