+++ src/main/java/com/rualone/app/domain/board/dto/response/PostDetailResponse.java
@@ -1,7 +1,7 @@
 package com.rualone.app.domain.board.dto.response;
 
-import com.rualone.app.domain.board.Comment;
-import com.rualone.app.domain.board.Post;
+import com.rualone.app.domain.board.entity.Comment;
+import com.rualone.app.domain.board.entity.Post;
 import lombok.Data;
 
 import java.util.ArrayList;
