+++ src/test/java/com/rualone/app/domain/board/application/PostServiceTest.java
@@ -1,6 +1,6 @@
 package com.rualone.app.domain.board.application;
 
-import com.rualone.app.domain.board.Post;
+import com.rualone.app.domain.board.entity.Post;
 import com.rualone.app.domain.board.dao.PostRepository;
 import com.rualone.app.domain.board.dto.request.PostCreateRequest;
 import org.junit.jupiter.api.AfterEach;
